refactor(wallet): break into multi files
