+++ wallet/wallet.go
@@ -2,35 +2,23 @@ package wallet
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/ptypes"
 	"github.com/op/go-logging"
-	"github.com/textileio/textile-go/crypto"
 	"github.com/textileio/textile-go/net"
-	nm "github.com/textileio/textile-go/net/model"
 	serv "github.com/textileio/textile-go/net/service"
-	"github.com/textileio/textile-go/pb"
 	trepo "github.com/textileio/textile-go/repo"
 	"github.com/textileio/textile-go/repo/db"
 	"github.com/textileio/textile-go/storage"
-	"github.com/textileio/textile-go/wallet/model"
 	"github.com/textileio/textile-go/wallet/thread"
 	"github.com/textileio/textile-go/wallet/util"
-	"gx/ipfs/QmSwZMWwFZSUpe5muU2xgTUwppH24KfMwdPXiwbEp2c6G5/go-libp2p-swarm"
 	"gx/ipfs/QmVW4cqbibru3hXA1iRmg85Fk7z9qML9k176CYQaMXVCrP/go-libp2p-kad-dht"
-	pstore "gx/ipfs/QmXauCuJzmzapetmC6W4TuDJLL1yFFrVzSHoWv8YdbmnxH/go-libp2p-peerstore"
-	libp2pn "gx/ipfs/QmXfkENeeBvh3zYA51MaSdGUdBjhQ99cP5WQe8zgr6wchG/go-libp2p-net"
-	mh "gx/ipfs/QmZyZDi491cCNTLfAhwcaDii2Kg4pwKRkhqQzURGDvY6ua/go-multihash"
 	libp2pc "gx/ipfs/QmaPbCnUMBohSGo3KnxEa2bHqyJVVeEEcwtqJAYxerieBo/go-libp2p-crypto"
 	utilmain "gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/cmd/ipfs/util"
 	oldcmds "gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/commands"
 	"gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/core"
 	"gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/repo/config"
 	"gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/repo/fsrepo"
-	uio "gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/unixfs/io"
 	"os"
 	"path/filepath"
 	"strings"
@@ -288,12 +276,12 @@ func (w *Wallet) Done() <-chan struct{} {
 	return w.done
 }
 
-func (w *Wallet) GetServerAddress() string {
-	return w.serverAddr
-}
-
 func (w *Wallet) GetRepoPath() string {
 	return w.repoPath
+}
+
+func (w *Wallet) GetServerAddress() string {
+	return w.serverAddr
 }
 
 // GetId returns peer id
@@ -352,414 +340,6 @@ func (w *Wallet) GetThread(id string) (*int, *thread.Thread) {
 	return nil, nil
 }
 
-// AddThread adds a thread with a given name and secret key
-func (w *Wallet) AddThread(name string, secret libp2pc.PrivKey) (*thread.Thread, error) {
-	skb, err := secret.Bytes()
-	if err != nil {
-		return nil, err
-	}
-	pkb, err := secret.GetPublic().Bytes()
-	if err != nil {
-		return nil, err
-	}
-	pk := libp2pc.ConfigEncodeKey(pkb)
-	threadModel := &trepo.Thread{
-		Id:      pk,
-		Name:    name,
-		PrivKey: skb,
-	}
-	if err := w.datastore.Threads().Add(threadModel); err != nil {
-		return nil, err
-	}
-
-	// load as active thread
-	thrd, err := w.loadThread(threadModel)
-	if err != nil {
-		return nil, err
-	}
-
-	// invite each device to the new thread
-	for _, device := range w.Devices() {
-		dpkb, err := libp2pc.ConfigDecodeKey(device.Id)
-		if err != nil {
-			return nil, err
-		}
-		dpk, err := libp2pc.UnmarshalPublicKey(dpkb)
-		if err != nil {
-			return nil, err
-		}
-		if _, err := thrd.AddInvite(dpk); err != nil {
-			return nil, err
-		}
-	}
-
-	// notify listeners
-	w.sendUpdate(Update{Id: thrd.Id, Name: thrd.Name, Type: ThreadAdded})
-
-	log.Debugf("added a new thread %s with name %s", thrd.Id, name)
-
-	return thrd, nil
-}
-
-// AddThreadWithMnemonic adds a thread with a given name and mnemonic phrase
-func (w *Wallet) AddThreadWithMnemonic(name string, mnemonic *string) (*thread.Thread, string, error) {
-	if mnemonic != nil {
-		log.Debugf("regenerating keypair from mnemonic for: %s", name)
-	} else {
-		log.Debugf("generating keypair for: %s", name)
-	}
-	secret, mnem, err := util.PrivKeyFromMnemonic(mnemonic)
-	if err != nil {
-		return nil, "", err
-	}
-	thrd, err := w.AddThread(name, secret)
-	if err != nil {
-		return nil, "", err
-	}
-	return thrd, mnem, nil
-}
-
-// RemoveThread removes a thread
-func (w *Wallet) RemoveThread(id string) (mh.Multihash, error) {
-	i, thrd := w.GetThread(id) // gets the loaded thread
-	if thrd == nil {
-		return nil, errors.New("thread not found")
-	}
-
-	// notify peers
-	addr, err := thrd.Leave()
-	if err != nil {
-		return nil, err
-	}
-
-	// remove model from db
-	if err := w.datastore.Threads().Delete(id); err != nil {
-		return nil, err
-	}
-
-	// clean up
-	thrd.Close()
-	copy(w.threads[*i:], w.threads[*i+1:])
-	w.threads[len(w.threads)-1] = nil
-	w.threads = w.threads[:len(w.threads)-1]
-
-	// notify listeners
-	w.sendUpdate(Update{Id: thrd.Id, Name: thrd.Name, Type: ThreadRemoved})
-
-	log.Infof("removed thread %s with name %s", id, thrd.Name)
-
-	return addr, nil
-}
-
-// AcceptThreadInvite attemps to download an encrypted thread key from an internal invite,
-// add the thread, and notify the inviter of the join
-func (w *Wallet) AcceptThreadInvite(blockId string) (mh.Multihash, error) {
-	// download
-	messageb, err := util.GetDataAtPath(w.ipfs, fmt.Sprintf("%s", blockId))
-	if err != nil {
-		return nil, err
-	}
-	message := new(pb.Message)
-	if err := proto.Unmarshal(messageb, message); err != nil {
-		return nil, err
-	}
-	signed := new(pb.SignedThreadBlock)
-	if err := ptypes.UnmarshalAny(message.Payload, signed); err != nil {
-		return nil, err
-	}
-	invite := new(pb.ThreadInvite)
-	if err := proto.Unmarshal(signed.Block, invite); err != nil {
-		return nil, err
-	}
-
-	// verify invitee
-	if invite.InviteeId != w.ipfs.Identity.Pretty() {
-		return nil, errors.New("invalid invitee")
-	}
-
-	// verify author sig
-	authorPk, err := libp2pc.UnmarshalPublicKey(invite.Header.AuthorPk)
-	if err != nil {
-		return nil, err
-	}
-	if err := crypto.Verify(authorPk, signed.Block, signed.AuthorSig); err != nil {
-		return nil, err
-	}
-
-	// decrypt thread key with private key
-	key, err := w.GetPrivKey()
-	if err != nil {
-		return nil, err
-	}
-	skb, err := crypto.Decrypt(key, invite.SkCipher)
-	if err != nil {
-		return nil, err
-	}
-	sk, err := libp2pc.UnmarshalPrivateKey(skb)
-	if err != nil {
-		return nil, err
-	}
-
-	// verify thread sig
-	if err := crypto.Verify(sk.GetPublic(), signed.Block, signed.ThreadSig); err != nil {
-		return nil, err
-	}
-
-	// add it
-	thrd, err := w.AddThread(invite.SuggestedName, sk)
-	if err != nil {
-		return nil, err
-	}
-
-	return thrd.Join(authorPk, blockId)
-}
-
-// AcceptExternalThreadInvite attemps to download an encrypted thread key from an external invite,
-// add the thread, and notify the inviter of the join
-func (w *Wallet) AcceptExternalThreadInvite(blockId string, key []byte) (mh.Multihash, error) {
-	// download
-	messageb, err := util.GetDataAtPath(w.ipfs, fmt.Sprintf("%s", blockId))
-	if err != nil {
-		return nil, err
-	}
-	message := new(pb.Message)
-	if err := proto.Unmarshal(messageb, message); err != nil {
-		return nil, err
-	}
-	signed := new(pb.SignedThreadBlock)
-	if err := ptypes.UnmarshalAny(message.Payload, signed); err != nil {
-		return nil, err
-	}
-	invite := new(pb.ThreadExternalInvite)
-	if err := proto.Unmarshal(signed.Block, invite); err != nil {
-		return nil, err
-	}
-
-	// verify author sig
-	authorPk, err := libp2pc.UnmarshalPublicKey(invite.Header.AuthorPk)
-	if err != nil {
-		return nil, err
-	}
-	if err := crypto.Verify(authorPk, signed.Block, signed.AuthorSig); err != nil {
-		return nil, err
-	}
-
-	// decrypt thread key
-	skb, err := crypto.DecryptAES(invite.SkCipher, key)
-	if err != nil {
-		return nil, err
-	}
-	sk, err := libp2pc.UnmarshalPrivateKey(skb)
-	if err != nil {
-		return nil, err
-	}
-
-	// verify thread sig
-	if err := crypto.Verify(sk.GetPublic(), signed.Block, signed.ThreadSig); err != nil {
-		return nil, err
-	}
-
-	// add it
-	thrd, err := w.AddThread(invite.SuggestedName, sk)
-	if err != nil {
-		return nil, err
-	}
-
-	return thrd.Join(authorPk, blockId)
-}
-
-// Devices lists all devices
-func (w *Wallet) Devices() []trepo.Device {
-	return w.datastore.Devices().List("")
-}
-
-// AddDevice creates an invite for every current and future thread
-func (w *Wallet) AddDevice(name string, pk libp2pc.PubKey) error {
-	if !w.Online() {
-		return ErrOffline
-	}
-
-	// index a new device
-	pkb, err := pk.Bytes()
-	if err != nil {
-		return err
-	}
-	deviceModel := &trepo.Device{
-		Id:   libp2pc.ConfigEncodeKey(pkb),
-		Name: name,
-	}
-	if err := w.datastore.Devices().Add(deviceModel); err != nil {
-		return err
-	}
-	log.Infof("added device '%s'", name)
-
-	// invite device to existing threads
-	for _, thrd := range w.threads {
-		if _, err := thrd.AddInvite(pk); err != nil {
-			return err
-		}
-	}
-
-	// notify listeners
-	w.sendUpdate(Update{Id: deviceModel.Id, Name: deviceModel.Name, Type: DeviceAdded})
-
-	return nil
-}
-
-// RemoveDevice removes a device
-func (w *Wallet) RemoveDevice(id string) error {
-	device := w.datastore.Devices().Get(id)
-	if device == nil {
-		return errors.New("device not found")
-	}
-	if err := w.datastore.Devices().Delete(id); err != nil {
-		return err
-	}
-	log.Infof("removed device '%s'", id)
-
-	// TODO: uninvite?
-
-	// notify listeners
-	w.sendUpdate(Update{Id: device.Id, Name: device.Name, Type: DeviceRemoved})
-
-	return nil
-}
-
-// AddPhoto add a photo to the local ipfs node
-func (w *Wallet) AddPhoto(path string) (*nm.AddResult, error) {
-	// get a key to encrypt with
-	key, err := crypto.GenerateAESKey()
-	if err != nil {
-		return nil, err
-	}
-
-	// read file from disk
-	file, err := os.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	// decode image
-	reader, format, err := util.DecodeImage(file)
-	if err != nil {
-		return nil, err
-	}
-
-	// make a thumbnail
-	reader.Seek(0, 0)
-	var thumbFormat util.ThumbnailFormat
-	if format == "gif" {
-		thumbFormat = util.GIF
-	} else {
-		thumbFormat = util.JPEG
-	}
-	thumb, err := util.MakeThumbnail(reader, thumbFormat, model.ThumbnailWidth)
-	if err != nil {
-		return nil, err
-	}
-
-	// get some meta data
-	username, _ := w.datastore.Profile().GetUsername() // ignore if not present (not signed in)
-	mpk, err := w.GetPubKey()
-	if err != nil {
-		return nil, err
-	}
-	mpkb, err := mpk.Bytes()
-	if err != nil {
-		return nil, err
-	}
-
-	// path info
-	fpath := file.Name()
-	ext := strings.ToLower(filepath.Ext(fpath))
-
-	// get metadata
-	reader.Seek(0, 0)
-	meta, err := util.GetMetadata(reader, fpath, ext, thumbFormat, username)
-	if err != nil {
-		return nil, err
-	}
-	metab, err := json.Marshal(meta)
-	if err != nil {
-		return nil, err
-	}
-
-	// encrypt files
-	reader.Seek(0, 0)
-	photocipher, err := util.GetEncryptedReaderBytes(reader, key)
-	if err != nil {
-		return nil, err
-	}
-	thumbcipher, err := crypto.EncryptAES(thumb, key)
-	if err != nil {
-		return nil, err
-	}
-	metacipher, err := crypto.EncryptAES(metab, key)
-	if err != nil {
-		return nil, err
-	}
-	mpkcipher, err := crypto.EncryptAES(mpkb, key)
-	if err != nil {
-		return nil, err
-	}
-
-	// create a virtual directory for the photo
-	dirb := uio.NewDirectory(w.ipfs.DAG)
-	err = util.AddFileToDirectory(w.ipfs, dirb, photocipher, "photo")
-	if err != nil {
-		return nil, err
-	}
-	err = util.AddFileToDirectory(w.ipfs, dirb, thumbcipher, "thumb")
-	if err != nil {
-		return nil, err
-	}
-	err = util.AddFileToDirectory(w.ipfs, dirb, metacipher, "meta")
-	if err != nil {
-		return nil, err
-	}
-	err = util.AddFileToDirectory(w.ipfs, dirb, mpkcipher, "pk")
-	if err != nil {
-		return nil, err
-	}
-
-	// pin the directory
-	dir, err := dirb.GetNode()
-	if err != nil {
-		return nil, err
-	}
-	if err := util.PinDirectory(w.ipfs, dir, []string{"photo"}); err != nil {
-		return nil, err
-	}
-	id := dir.Cid().Hash().B58String()
-
-	// create and init a new multipart request
-	request := &net.PinRequest{}
-	request.Init(filepath.Join(w.repoPath, "tmp"), id)
-
-	// add files to request
-	if err := request.AddFile(photocipher, "photo"); err != nil {
-		return nil, err
-	}
-	if err := request.AddFile(thumbcipher, "thumb"); err != nil {
-		return nil, err
-	}
-	if err := request.AddFile(metacipher, "meta"); err != nil {
-		return nil, err
-	}
-	if err := request.AddFile(mpkcipher, "pk"); err != nil {
-		return nil, err
-	}
-
-	// finish request
-	if err := request.Finish(); err != nil {
-		return nil, err
-	}
-
-	// all done
-	return &nm.AddResult{Id: id, Key: string(key), PinRequest: request}, nil
-}
-
 // GetBlock searches for a local block associated with the given target
 func (w *Wallet) GetBlock(id string) (*trepo.Block, error) {
 	block := w.datastore.Blocks().Get(id)
@@ -784,107 +364,6 @@ func (w *Wallet) GetDataAtPath(path string) ([]byte, error) {
 		return nil, ErrStopped
 	}
 	return util.GetDataAtPath(w.ipfs, path)
-}
-
-// ConnectPeer connect to another ipfs peer (i.e., ipfs swarm connect)
-func (w *Wallet) ConnectPeer(addrs []string) ([]string, error) {
-	if !w.Online() {
-		return nil, ErrOffline
-	}
-	snet, ok := w.ipfs.PeerHost.Network().(*swarm.Network)
-	if !ok {
-		return nil, errors.New("peerhost network was not swarm")
-	}
-
-	swrm := snet.Swarm()
-	pis, err := util.PeersWithAddresses(addrs)
-	if err != nil {
-		return nil, err
-	}
-
-	output := make([]string, len(pis))
-	for i, pi := range pis {
-		swrm.Backoff().Clear(pi.ID)
-
-		output[i] = "connect " + pi.ID.Pretty()
-
-		err := w.ipfs.PeerHost.Connect(w.ipfs.Context(), pi)
-		if err != nil {
-			return nil, fmt.Errorf("%s failure: %s", output[i], err)
-		}
-		output[i] += " success"
-	}
-	return output, nil
-}
-
-// PingPeer pings a peer num times, returning the result to out chan
-func (w *Wallet) PingPeer(addrs string, num int, out chan string) error {
-	if !w.started {
-		return ErrStopped
-	}
-	if !w.Online() {
-		return ErrOffline
-	}
-	addr, pid, err := util.ParsePeerParam(addrs)
-	if addr != nil {
-		w.ipfs.Peerstore.AddAddr(pid, addr, pstore.TempAddrTTL) // temporary
-	}
-
-	if len(w.ipfs.Peerstore.Addrs(pid)) == 0 {
-		// Make sure we can find the node in question
-		log.Debugf("looking up peer: %s", pid.Pretty())
-
-		ctx, cancel := context.WithTimeout(w.ipfs.Context(), pingTimeout)
-		defer cancel()
-		p, err := w.ipfs.Routing.FindPeer(ctx, pid)
-		if err != nil {
-			err = fmt.Errorf("peer lookup error: %s", err)
-			log.Errorf(err.Error())
-			return err
-		}
-		w.ipfs.Peerstore.AddAddrs(p.ID, p.Addrs, pstore.TempAddrTTL)
-	}
-
-	ctx, cancel := context.WithTimeout(w.ipfs.Context(), pingTimeout*time.Duration(num))
-	defer cancel()
-	pings, err := w.ipfs.Ping.Ping(ctx, pid)
-	if err != nil {
-		log.Errorf("error pinging peer %s: %s", pid.Pretty(), err)
-		return err
-	}
-
-	var done bool
-	var total time.Duration
-	for i := 0; i < num && !done; i++ {
-		select {
-		case <-ctx.Done():
-			done = true
-			close(out)
-			break
-		case t, ok := <-pings:
-			if !ok {
-				done = true
-				close(out)
-				break
-			}
-			total += t
-			msg := fmt.Sprintf("ping %s completed after %f seconds", pid.Pretty(), t.Seconds())
-			select {
-			case out <- msg:
-			default:
-			}
-			log.Debug(msg)
-			time.Sleep(time.Second)
-		}
-	}
-	return nil
-}
-
-func (w *Wallet) Peers() ([]libp2pn.Conn, error) {
-	if !w.Online() {
-		return nil, ErrOffline
-	}
-	return w.ipfs.PeerHost.Network().Conns(), nil
 }
 
 // createIPFS creates an IPFS node
